requirements.txt
Flask
Werkzeug
piexif
Pillow
flask_fontawesome
